Makefile: initialize Makefile
A generic Makefile has been set up. This makes it easier to compile the
program over and over again, and also does not require any major changes
to be made to the Makefile everytime something new is created. This makes
it easier to manage dependencies in the Makefile, and everything else too,
overall.

Only when a new source file or header file is created, the Makefile needs
to be updated. The places where they need to be updated have been
highlighted using comments.

This make file preserves the directory structure of the src/ directory,
i.e. object files are generated in bin/ in the same directory structure
as observed in src/.
Note: If a directory that is supposed to be housing the object files
being generated doesn't already exist, then it is automatically created
by the Makefile.

The final executable image obtained is "vEB", and is created in the root
of the repository.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>