rb_iterator, rb_tree: declare classes RBTree, and Iterator in it
Set up the Iterator class, and basic iterator functions for iteration
through RBNodes in a given RBTree.

The Iterator class is nested in the RBTree class, so we can create as
many RBTree::Iterators as we want for a given tree.

Note: The RBTree's class layout itself is yet to made concrete.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>