rb_tree: set up default constructor
Set up the default constructor for RBTree.

Modify the Color values of 'Black' and 'Red' to 'BLACK' and 'RED'
respectively for better readability.

The default constructor for RBNode need not be deleted. It can be used
to instantiate an uninitialized object of type RBNode instead. This might
come in handy at some point in the future, and if it doesn't, then it can
just as easily be deleted.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>