rb_tree: insert: handle case where insertion is supposed to fail
If a value that already exists in the RBT is required to be added, then
ignore it, and don't allow insertion to occur in that case.

Only if RBT_UNIQUE is defined, will the RBTree not allow duplicates
values. Else, duplicates will be allowed. This may probably have to be
reversed too, depending upon how further implementations turn out.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>