utils.h: set up pretty printing, better display function definitions
* Set up pretty printing for the trees created. Calling "print_tree"
will print the tree in a nice way. This is likely to be replaced by
operator overloading.

* Rename the preorder, inorder, and postorder display functions. Also,
move out the actual functions into utils.h, and provide  the
class with wrappers instead. This may also be made entirely independent,
and the member functions may be removed from the RBTree class entirely.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>