pretty printing: properly set up pretty printing for RBTree using "<<"
* Initialize the "<<" operator as a friend function, and define the same.

* Rename print_binary_tree() to print_tree(). Update the parameters
passed to this function to include an ostream& variable as well, so it
fits correctly with the call being made from the operator<< function.
Accordingly, use ostream& variable in place of std::cout.

* Update the RBTree() constructor. This is supposed to be creating an empty
RBTree. However, previously it also wrongly initialized NIL to point to
a new instance of RBNode. NIL must always = nullptr. Fix this issue.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>