rb_tree.h: fix insert()
The insert function seems to give segmentation faults everytime a skewed
tree is created (e.g., if nodes were inserted in the order - 11, 4, 2).

Turns out, these issues were majorly being created since leaf nodes would
point to nullptr instead of NIL(and even if they did point to NIL, NIL
was a nullptr, so there was no real effect).
Making NIL actually a node (albeit an empty one), and rewriting the insert
function appropriately fixes this issue.

Modify the RBNode constructors, so that default constructor is in play,
and if not, atleast the value of the node has to be passed to the ctor.

Remove tree_insert(), and rewrite insert() in line with the actually
correct, true CLRS explanation. insert() now uses the helper function
insert_fixup() instead.

Fixes: 29c4c56 ("RBTree: define the basic insert() function")
Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>