RBTree: support predicates and predicate-based ordering
Introduce support for predicates and predicate-based ordering of nodes in
Red-Black trees.

This is not necessary for RBNode, since RBNode itself does not perform
any sort of predicate-based operations. However, this should be
introduced for RBTree, since RBTree must decide using what predicate
the RBNodes in the RBTree are ordered.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>