rb_tree.h: add condition in delete_fixup()
In the case where the RBTree has only 3 nodes left, when delete_fixup()
is called, it will result in a segfault, as the nodes NIL
refers to (which are nullptrs) are accessed.

Fix this issue by adding a clause in the while loop that checks if the
node is NIL or not before proceeding.

Fixes: eb24260 ("rb_tree.h: fix delete - deletion of a single node")
Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>