rb_tree.h: handle NIL nodes correctly in delete_fixup()
Commit e6f085b handled NIL nodes to prevent segfaults. While this prevented
segfaults, it also messed up the coloring of the tree, which was also
supposed to be fixed up. Further, the wrong node was being checked for
NIL, which just made things worse.
There was also an assignment being done incorrectly instead of a
comparison within the initial if condition in the while loop.

Update the check for NIL node better within the loop itself by
identifying the actual node which shouldn't be NIL, and enforcing a
check for the same. If the node ('w') is found to be NIL, then break out
of the while loop within delete_fixup().

Fixes: e6f085b ("rb_tree.h: add condition in delete_fixup()")
Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>