RBNode: declare all data members as private members
* Declare all data members of class RBNode to be private.
* Add RBTree as a friend class, and the functions in utils.h as
friend functions, so nothing breaks.
* Make necessary changes in the other files as well, so they can still
access RBNode's newly privatized members.

For Iterator, the << operator function that returns an ostream reference
itself is a friend function of the Iterator class. Since RBTree is
already a friend class and the Iterator class is nested in RBTree, the
Iterator class itself has access to the private objects of RBNode.
So, create a new display() function in Iterator, which will write to a
stream, and will default to writing into std::cout.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>