NIL: fix memory leak caused by NIL
NIL is a static pointer that points to dynamically allocated memory.
So, if left as-is, it leads to a memory leak. Converting NIL to a smart
pointer - unique_ptr, fixes this memory leak.

Note:
An atexit() function could be declared in the constructor too as an
alternative. However, this is less favorable, since the atexit()
function will get registered multiple times needlessly.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>